user-service/requirements.txt
flask
pymongo
werkzeug
pyjwt
